pyproject.toml
[project]
name = "birds-of-play"
version = "0.1.0"
description = "Motion detection library with Python bindings"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Birds of Play Team", email = "team@birdsofplay.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Computer Vision",
]
requires-python = ">=3.8"
dependencies = [
    "pybind11>=3.0.0",
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "pymongo>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
]
ml = [
    # Core ML dependencies
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "torch>=1.12.0",
    "torchvision>=0.13.0",
    # Visualization
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "plotly>=5.0.0",
    "networkx>=2.8.0",
    # Dimensionality reduction
    "umap-learn>=0.5.0",
    # Image processing
    "Pillow>=8.3.0",
    # Web interface
    "flask>=2.0.0",
    "flask-cors>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/louisbove84/birds_of_play"
Repository = "https://github.com/louisbove84/birds_of_play.git"
Documentation = "https://github.com/louisbove84/birds_of_play#readme"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
    ".venv",
    "venv",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
